Add files via upload
The given dataset contains details about organic chemical compounds including their chemical features, isomeric conformation, names and the classes in which they are classified. The compounds are classified as either ‘Musk’ or ‘Non-Musk’ compounds. The task is to make a classification model where I have used DNN and build a model and got the accuracy of  0.9962121.