Add files via upload
The maximum of the data is not similar but found that row 2284 and 3403 cosine similarity close to 1.  If we want to use a neural network with this, we can use come pre-trained corpus module to predict a word give a context. (Bag of words model). Some of the pre-trained word embeddings can be found from google.  We can use those for comparing text 1 and text 2 using the genism library, the n-similarity function which uses cosine similarity. Cosine similarity is been used directly I  have not used the genism library which can be used to if we consider it to be unsupervised problem.